[PM-1284] Pr comments
Split DynamicTlsPeerGroup into smaller files
Use separate library to handle availabity interval
Extract help methofs for keyfactory and ecdsa signature
Fix few typos
