+++ scalanet/src/io/iohk/scalanet/peergroup/InetMultiAddress.scala
@@ -47,7 +47,7 @@ case class InetMultiAddress(inetSocketAddress: InetSocketAddress) {
 }
 
 object InetMultiAddress {
-  implicit val adressInst = new Addressable[InetMultiAddress] {
+  implicit val addressableInetMultiAddressInst = new Addressable[InetMultiAddress] {
     override def getAddress(a: InetMultiAddress): InetSocketAddress = a.inetSocketAddress
   }
 }
